Create separate sync feature in core
