+++ app/src/main/java/com/holparb/notemark/core/datasync/data/NoteSyncWorker.kt
@@ -1,18 +1,18 @@
-package com.holparb.notemark.notes.data.sync
+package com.holparb.notemark.core.datasync.data
 
 import android.content.Context
 import androidx.work.CoroutineWorker
 import androidx.work.WorkerParameters
-import com.holparb.notemark.notes.domain.repository.NoteRepository
+import com.holparb.notemark.core.datasync.domain.DataSyncRepository
 
 class NoteSyncWorker (
-    private val noteRepository: NoteRepository,
+    private val dataSyncRepository: DataSyncRepository,
     appContext: Context,
     params: WorkerParameters
 ): CoroutineWorker(appContext = appContext, params = params) {
 
     override suspend fun doWork(): Result {
-        val syncResult = noteRepository.syncNotes()
+        val syncResult = dataSyncRepository.syncNotes()
         return when(syncResult) {
             is com.holparb.notemark.core.domain.result.Result.Error -> Result.retry()
             is com.holparb.notemark.core.domain.result.Result.Success -> Result.success()
