+++ app/src/main/java/com/holparb/notemark/notes/domain/repository/NoteRepository.kt
@@ -7,7 +7,6 @@ import kotlinx.coroutines.flow.Flow
 
 interface NoteRepository {
 
-    suspend fun getNotes(page: Int = -1, size: Int = 20): Result<Unit, DataError>
     fun observeNotes(): Result<Flow<List<Note>>, DataError.LocalError>
     suspend fun getNote(noteId: String): Result<Note, DataError.LocalError>
     suspend fun createNewNoteInDatabase(): Result<String, DataError.LocalError>
@@ -15,5 +14,4 @@ interface NoteRepository {
     suspend fun updateNote(note: Note): Result<Unit, DataError>
     suspend fun deleteNote(noteId: String): Result<Unit, DataError>
     suspend fun deleteNoteFromDatabase(noteId: String): Result<Unit, DataError.LocalError>
-    suspend fun syncNotes(): Result<Unit, DataError>
 }
